backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
sqlalchemy
pymysql
opencv-python
Pillow
